+++ hart/constants.py
@@ -1,4 +1,5 @@
 DEBIAN_VERSIONS = {
+    'bookworm': 12,
     'bullseye': 11,
     'buster': 10,
     'stretch': 9,