+++ hart/constants.py
@@ -1,7 +1,5 @@
 DEBIAN_VERSIONS = {
-    'bookworm': 12,
-    'bullseye': 11,
-    'buster': 10,
-    'stretch': 9,
-    'jessie': 8,
+    "bookworm": 12,
+    "bullseye": 11,
+    "buster": 10,
 }